+++ app/page.tsx
@@ -1,4 +1,5 @@
 import { Certifications } from "./_components/Certifications";
+import { Contact } from "./_components/Contact";
 import { Header } from "./_components/Header";
 import { Hero } from "./_components/Hero";
 import { Skills } from "./_components/Skills";
@@ -17,6 +18,8 @@ export default function Home() {
         <Skills />
         <Spacing size="sm" />
         <Certifications />
+        <Spacing size="sm" />
+        <Contact />
         <Spacing size="sm" />
       </div>
     </main>